bigquery/requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
elasticsearch==6.1.1
futures==3.2.0
google-api-core==1.3.0
google-auth-oauthlib==0.2.0
google-auth==1.5.1
google-cloud-bigquery==1.5.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.7
./indexer_util
jsmin==2.2.2
numpy==1.15.1
oauthlib==2.1.0
pandas-gbq==0.3.1
pandas==0.22.0
protobuf==3.6.1
pyasn1-modules==0.2.2
pyasn1==0.4.4
python-dateutil==2.7.3
pytz==2018.5
requests-oauthlib==1.0.0
requests==2.19.1
rsa==3.4.2
six==1.11.0
urllib3==1.22
